+++ GitHub/Auth/Device-Authorization-Grant/Package.swift
@@ -1,0 +1,19 @@
+// swift-tools-version: 5.7
+
+import PackageDescription
+
+let package = Package(
+    name: "auth-device",
+    products: [
+        .executable(name: "auth-device", targets: ["auth-device"])
+    ],
+    dependencies: [
+    ],
+    targets: [
+        .executableTarget(
+            name: "auth-device",
+            dependencies: [
+            ],
+            path: "./src")
+    ]
+)