+++ HISTORY.md
@@ -5,3 +5,4 @@
  - 2022-10-06 - MyIP: add display public IP address macOS Monterey SwiftUI program
  - 2022-10-24 - SwiftRequest: displays details on the HTTP request, container environment and directory contents for a Swift container deployed to Google Cloud Run
  - 2022-11-27 - Gist-List: duplicates the GitHub CLI command: `gh gist list`
+ - 2022-11-29 - Auth-Device: GitHub Device Authorization Grant